// frontOffice/src/app/core/models/form-field-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface FormFieldDto {
  javaRegex?: string;
  label: string;
  order?: number;
  placeholder?: string;
  rangeEnd?: number;
  rangeStart?: number;
  rangeValid?: boolean;
  regex?: string;
  regexErrorMessage?: string;
  required?: boolean;
  selectOptions?: Array<string>;
  type: string;
}