// frontOffice/src/app/core/models/auto-insurance-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AddressInfo } from '../models/address-info';
export interface AutoInsuranceRequest {
  accidentHistory?: string;
  addressInfo?: AddressInfo;
  billingPeriod?: 'MONTHLY' | 'QUARTERLY' | 'SEMI_ANNUAL' | 'ANNUAL';
  coverageType?: string;
  defensiveDrivingCourse?: boolean;
  drivingExperience?: number;
  licenseNumber?: string;
  trafficViolations?: boolean;
  vehicleMake?: string;
  vehicleModel?: string;
  vehicleType?: string;
  vin?: string;
}